k8s-learning-app/requirements.txt
Flask==3.0.0
Markdown==3.5.1
PyYAML==6.0.1
Werkzeug==3.0.1
